Fix: [BADA-121] trade 파일구조로 바꾸기 #65
