--- src/features/mypage/model/sosHistory.ts
@@ -1,7 +1,0 @@
-export interface SosHistory {
-  createdAt: string;
-  merchantName: string;
-  menuName: string;
-  price: number;
-  imageUrl: string;
-}
