+++ src/features/mypage/model/useSosHistoryQuery.ts
@@ -1,0 +1,9 @@
+import { useQuery } from '@tanstack/react-query';
+import { getSosHistory } from '../api/mypage';
+
+export const useSosHistoryQuery = () => {
+  return useQuery({
+    queryKey: ['sosHistory'],
+    queryFn: getSosHistory,
+  });
+};
