+++ src/features/mypage/model/useUserCoinQuery.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
-import { getUserCoin } from '../api/userCoin';
-import { UserCoin } from '../model/userCoin';
+import { getUserCoin } from '../api/mypage';
+import { UserCoin } from '../lib/types';
 
 export const useUserCoin = () => {
   return useQuery<UserCoin>({
