+++ src/features/mypage/model/useUserDataUsageQuery.ts
@@ -1,12 +1,10 @@
 import { useQuery } from '@tanstack/react-query';
-import { getUserDataUsage } from '../api/userDataUsage';
-import { UserDataUsage } from '../model/userDataUsage';
+import { getUserDataUsage } from '../api/mypage';
+import { UserDataUsage } from '../lib/types';
 
 export const useUserDataUsage = () => {
   return useQuery<UserDataUsage>({
     queryKey: ['userDataUsage'],
     queryFn: getUserDataUsage,
-    staleTime: 1000 * 60,
-    refetchOnWindowFocus: false,
   });
 };
