--- src/features/mypage/model/userCoin.ts
@@ -1,3 +1,0 @@
-export interface UserCoin {
-  coin: number;
-}
