--- src/features/mypage/model/userDataUsage.ts
@@ -1,3 +1,0 @@
-export interface UserDataUsage {
-  used: number;
-}
