+++ src/features/mypage/pages/MySosPage.tsx
@@ -1,17 +1,17 @@
 'use client';
 
-import { useSosHistory } from '../hooks/useSosHistory';
-import { SosHistory } from '../model/sosHistory';
+import { SosHistoryItem } from '../lib/types';
+import { useSosHistoryQuery } from '../model/useSosHistoryQuery';
 
 export const SosHistoryList = () => {
-  const { data, isLoading, isError } = useSosHistory();
+  const { data, isLoading, isError } = useSosHistoryQuery();
 
   if (isLoading) return <p>로딩 중...</p>;
   if (isError || !data) return <p>요청 내역을 불러오지 못했습니다.</p>;
 
   return (
     <div className="flex flex-col gap-4">
-      {data.map((item: SosHistory, i: number) => (
+      {data.map((item: SosHistoryItem, i: number) => (
         <div key={i} className="flex items-start gap-4">
           <img src={item.imageUrl} alt="상품 이미지" className="w-20 h-20 rounded-md bg-gray-200" />
           <div className="flex flex-col justify-between flex-1">
