+++ src/shared/lib/axios/axiosInstance.ts
@@ -1,5 +1,5 @@
-import axios from 'axios';
 import { BASE_URL, NETWORK_TIMEOUT } from '@shared/constants/api';
+import axios from 'axios';
 import { applyInterceptors } from './axiosInterceptor';
 
 export const axiosInstance = axios.create({
@@ -10,10 +10,3 @@ export const axiosInstance = axios.create({
 
 applyInterceptors(axiosInstance);
 
-axiosInstance.interceptors.request.use((config) => {
-  const token = localStorage.getItem('accessToken');
-  if (token) {
-    config.headers.Authorization = `Bearer ${token}`;
-  }
-  return config;
-});
